+++ include/utils/json.h
@@ -2,6 +2,6 @@
 #define JSON_H
 
 char* get_json_value(const char* src, const char* key);
-char** separate_array(char* src, int* num_elems, int capacity);
+char** separate_array(char* src, size_t* num_elems, size_t capacity);
 
 #endif
