+++ include/utils/utils.h
@@ -7,10 +7,10 @@
 char* skip_whitespace(char* src);
 
 // Returns len of url-decoded string
-int url_decoded_str_len(char* src);
+size_t url_decoded_str_len(char* src);
 
 // Decodes url-encoded string
-void decode_url(char* dst, char* src, int src_len);
+void decode_url(char* dst, char* src, size_t src_len);
 
 // String comparison function with the option for case sensitivity
 bool str_equals(const char* a, const char* b, bool case_sensitive);
